+++ tests/eurotherm_eibisynch.py
@@ -1,8 +1,12 @@
 import unittest
 
+from parameterized import parameterized
+
 from utils.test_modes import TestModes
 from utils.ioc_launcher import get_default_ioc_dir, ProcServLauncher
-from common_tests.eurotherm import EurothermBaseTests
+from common_tests.eurotherm import (EurothermBaseTests, NONE_TXT_CALIBRATION_MIN_TEMPERATURE,
+                                    NONE_TXT_CALIBRATION_MAX_TEMPERATURE)
+from utils.calibration_utils import use_calibration_file
 
 # Internal Address of device (must be 2 characters)
 ADDRESS = "A01"
@@ -47,3 +51,51 @@ class EurothermTests(EurothermBaseTests, unittest.TestCase):
 
     def get_emulator_device(self):
         return EMULATOR_DEVICE
+
+    @parameterized.expand([
+        ("over_range_calc_pv_is_over_range", NONE_TXT_CALIBRATION_MAX_TEMPERATURE + 5.0, 1.0),
+        ("over_range_calc_pv_is_within_range", NONE_TXT_CALIBRATION_MAX_TEMPERATURE - 200, 0.0),
+        ("over_range_calc_pv_is_within_range", NONE_TXT_CALIBRATION_MAX_TEMPERATURE, 0.0)
+    ])
+    def test_GIVEN_None_txt_calibration_file_WHEN_temperature_is_set_THEN(
+            self, _, temperature, expected_value_of_over_range_calc_pv):
+        """
+        Note: this test can only run on BISYNCH eurotherms, modbus max temperature is 6553.5 but ramp file goes up
+        to 10,000 and this test attempts to check this behaviour
+        """
+        # Arrange
+
+        self._assert_using_mock_table_location()
+        with use_calibration_file(self.ca, "None.txt"):
+            self.ca.assert_that_pv_exists("CAL:RANGE")
+            self.ca.assert_that_pv_is("TEMP:RANGE:OVER.B", NONE_TXT_CALIBRATION_MAX_TEMPERATURE)
+
+            # Act:
+            self._set_setpoint_and_current_temperature(temperature)
+
+            # Assert
+            self.ca.assert_that_pv_is("TEMP:RANGE:OVER.A", temperature)
+            self.ca.assert_that_pv_is("TEMP:RANGE:OVER", expected_value_of_over_range_calc_pv)
+
+    def test_GIVEN_None_txt_calibration_file_WHEN_changed_to_C006_txt_calibration_file_THEN_the_calibration_limits_change(
+            self):
+        """
+        Note: this test can only run on BISYNCH eurotherms, modbus max temperature is 6553.5 but ramp file goes up
+        to 10,000 and this test attempts to check this behaviour
+        """
+        C006_CALIBRATION_FILE_MAX = 330.26135292267900000000
+        C006_CALIBRATION_FILE_MIN = 1.20927230303971000000
+
+        # Arrange
+        self._assert_using_mock_table_location()
+        with use_calibration_file(self.ca, "None.txt"):
+            self.ca.assert_that_pv_exists("CAL:RANGE")
+            self.ca.assert_that_pv_is("TEMP:RANGE:OVER.B", NONE_TXT_CALIBRATION_MAX_TEMPERATURE)
+            self.ca.assert_that_pv_is("TEMP:RANGE:UNDER.B", NONE_TXT_CALIBRATION_MIN_TEMPERATURE)
+
+        # Act:
+        with use_calibration_file(self.ca, "C006.txt"):
+
+            # Assert
+            self.ca.assert_that_pv_is("TEMP:RANGE:OVER.B", C006_CALIBRATION_FILE_MAX)
+            self.ca.assert_that_pv_is("TEMP:RANGE:UNDER.B", C006_CALIBRATION_FILE_MIN)
